perf(ci): split and shard the `agents-host` debug-view tests
`agents-host` arrived with #5988 and is the most expensive package in the
workspace, but it has no entry in `WORKSPACE_TEST_WEIGHTS`, so the balancer
placed it at the default weight of 1 -- as though it were among the cheapest.
Whichever shard received it ran about seven times as long as its siblings: the
`Test (3/8)` job has been taking around fourteen minutes against a calibration
of two, with `agents-host` alone accounting for twelve of them.

Weighting it honestly was not enough on its own, because the package's cost sat
in a single file. `debug_view_deployment_test.ts` was 77s of the package's 90s
locally -- sixteen tests, each standing up runtimes and deploying views, none
of them padded with sleeps or polls -- and the sharder distributes whole files,
so no arrangement could split that file's cost.

So the file is now three, along the seam its own test names already described:
pattern behavior, deployment lifecycle, and view replacement. Their shared
fixtures move to `debug_view_support.ts`, which is not itself a test module.
The three run at 40s, 33s and 16s where one ran at 77s, and the set of test
names is unchanged.

With that seam available, the package joins `cli`, `piece` and `tasks` in
sharding internally, at three ways. Its shards measure 37s, 30s and 14s, and
carry 6, 8 and 50 tests -- 64 in total, as before.

The modeled two-worker makespan rises from 73.6 to 79.1 against a ceiling of
80. That ceiling is now respected on an honest model rather than on a placement
that assumed this package was free. The weights here are local seconds, which
understates continuous integration, so the remaining headroom is thinner than
the number suggests.
