+++ packages/agents-host/deno.jsonc
@@ -2,7 +2,7 @@
   "name": "@commonfabric/agents-host",
   "tasks": {
     "start": "deno run -A ./main.ts",
-    "test": "deno test --allow-env --allow-read --allow-write --allow-run --allow-ffi --allow-sys=homedir,uid"
+    "test": "deno run --allow-env --allow-read --allow-run=$(deno eval \"console.log(Deno.execPath())\") ../../tasks/run-sharded-test-files.ts AGENTS_HOST_TEST_SHARD agents-host . -- --no-check --allow-env --allow-read --allow-write --allow-run --allow-ffi --allow-sys=homedir,uid"
   },
   "exports": {
     ".": "./mod.ts",
